components: if we're playing, we ain't offline
